Adds oscillate function to the world
