+++ lib/index.js
@@ -8,8 +8,8 @@ const World = require('./world');
 const HealthFood = require('./health-food');
 const RandomHealthFood = require('./random-health-food');
 // const Overlord = require('./overlord');
-const JunkFood = require('./junk-food');
-const RandomJunkFood = require('./random-junk-food');
+// const JunkFood = require('./junk-food');
+// const RandomJunkFood = require('./random-junk-food');
 // const Gallery = require('./gallery');
 // const ScoreBoard = require('./score-board');
 // const TrayCollision = require('./tray-collision');
@@ -49,7 +49,7 @@ startButton.addEventListener('click', function(){
   world.makeNewGame();
   masterGameLoop();
   world.userInput();
-  increaseDifficulty(overlord, world);
+  world.increaseDifficulty(overlord, world);
 });
   
 // function userInput(){
@@ -60,7 +60,7 @@ startButton.addEventListener('click', function(){
 // }  
 function masterGameLoop(){
   requestAnimationFrame(function gameLoop(){
-    checkGameEnders();
+    world.checkGameEnders();
     beginGame();
     trayCollision.checkLevel();
   reqAnimFrameID = requestAnimationFrame(gameLoop);
@@ -91,26 +91,26 @@ function masterGameLoop(){
 // }
 
 //pull this into a level.js or game.js
-function increaseDifficulty(overlord, world){
-  var interval = 500;
-  function oscillate(overlord, world){
-    interval = interval - 2;
-    if(overlord.x < 0 || overlord.x > world.width - overlord.width){
-      overlord.speed = -overlord.speed;
-    }
-    overlord.x += overlord.speed;
-
-    if(overlord.x%(Math.floor(Math.random() * 100)) === 0){
-      createJunkFood(overlord);
-    }
-    if (interval >= 50){
-      window.setTimeout(oscillate, interval, overlord, world);
-    } else if (interval < 75){
-      window.setTimeout(oscillate, 75, overlord, world);
-    }
-  }
-  oscillate(overlord, world);
-}
+// function increaseDifficulty(overlord, world){
+//   var interval = 500;
+//   function oscillate(overlord, world){
+//     interval = interval - 2;
+//     if(overlord.x < 0 || overlord.x > world.width - overlord.width){
+//       overlord.speed = -overlord.speed;
+//     }
+//     overlord.x += overlord.speed;
+// 
+//     if(overlord.x%(Math.floor(Math.random() * 100)) === 0){
+//       createJunkFood(overlord);
+//     }
+//     if (interval >= 50){
+//       window.setTimeout(oscillate, interval, overlord, world);
+//     } else if (interval < 75){
+//       window.setTimeout(oscillate, 75, overlord, world);
+//     }
+//   }
+//   oscillate(overlord, world);
+// }
 //pull this into level or game
 // function checkLevel(){
 //   var level = document.getElementById('level');
@@ -120,56 +120,56 @@ function increaseDifficulty(overlord, world){
 //     level.innerHTML = "<h3>Level 3</h3>";
 //   }
 // }
-function checkGameEnders(){
-  if(trayCollision.total < 0 && gameRunning || trayCollision.total >= 600 && gameRunning || !gameRunning){
-    endGame();
-  }
-}
+// function checkGameEnders(){
+//   if(trayCollision.total < 0 && gameRunning || trayCollision.total >= 600 && gameRunning || !gameRunning){
+//     endGame();
+//   }
+// }
 
 // //refactor into functions but mostly stay here
-function endGame(){
-  if(trayCollision.total >= 600){
-    addScore(trayCollision.total);
-    alert("You won! Your score was " + trayCollision.total + " points.");
-    endGameStyling();
-  } else if (trayCollision.total < 0){
-    alert("You lost! Eat more vegetables next time to defeat the Empty Calorie Overlord.");
-    endGameStyling();
-  }
-  stopGameMovement();
-  resetGame();
-  document.getElementById('score-game').innerHTML = "Score: " + trayCollision.total;
-  return gameRunning = false;
-}
+// function endGame(){
+//   if(trayCollision.total >= 600){
+//     addScore(trayCollision.total);
+//     alert("You won! Your score was " + trayCollision.total + " points.");
+//     endGameStyling();
+//   } else if (trayCollision.total < 0){
+//     alert("You lost! Eat more vegetables next time to defeat the Empty Calorie Overlord.");
+//     endGameStyling();
+//   }
+//   stopGameMovement();
+//   resetGame();
+//   document.getElementById('score-game').innerHTML = "Score: " + trayCollision.total;
+//   return gameRunning = false;
+// }
 
-function resetGame(){
-  healthFoods = [];
-  junkFoods = [];
-  trayCollision.total = 0;
-}
+// function resetGame(){
+//   healthFoods = [];
+//   junkFoods = [];
+//   trayCollision.total = 0;
+// }
 
-function endGameStyling(){
-  document.getElementById('instructions').style = "display: block";
-  document.getElementById('start-btn').style = "display: block";
-  document.getElementById('score-title').style = "display: none";
-  document.getElementById('high-scores').style = "display: none";
-  canvas.style.display = 'none';
-  startButton.style.display = 'block';
-}
+// function endGameStyling(){
+//   document.getElementById('instructions').style = "display: block";
+//   document.getElementById('start-btn').style = "display: block";
+//   document.getElementById('score-title').style = "display: none";
+//   document.getElementById('high-scores').style = "display: none";
+//   canvas.style.display = 'none';
+//   startButton.style.display = 'block';
+// }
 
-function stopGameMovement(){
-  for (var i = 0; i < 1000; i++) {
-    window.clearInterval(i);
-  }
-}
+// function stopGameMovement(){
+//   for (var i = 0; i < 1000; i++) {
+//     window.clearInterval(i);
+//   }
+// }
 //add this to score board
-function addScore(winningScore){
-  if(scoreBoard.scores[0] && trayCollision.total > Math.min.apply(Math, scoreBoard.scores)){
-    scoreBoard.addHighScoreToPage(winningScore);
-  } else {
-    scoreBoard.addHighScoreToPage(winningScore);
-  }
-}
+// function addScore(winningScore){
+//   if(scoreBoard.scores[0] && trayCollision.total > Math.min.apply(Math, scoreBoard.scores)){
+//     scoreBoard.addHighScoreToPage(winningScore);
+//   } else {
+//     scoreBoard.addHighScoreToPage(winningScore);
+//   }
+// }
 
 function beginGame(){
   world.draw();
@@ -211,10 +211,10 @@ function foodDraw(food){
   ctx.drawImage(gallery.images[foodImage], food.x, food.y, food.width, food.height);
 }
 //move to junk food or overlord depending (probably overlord)
-function createJunkFood(overlord){
-  var overlordMouth = Math.floor(overlord.x + (overlord.width/2));
-  var randomJunkFoodAttributes = new RandomJunkFood(overlordMouth);
-  var junkFood = new JunkFood(randomJunkFoodAttributes);
-  junkFood.ctx = ctx;
-  junkFoods.push(junkFood);
-}
+// function createJunkFood(overlord){
+//   var overlordMouth = Math.floor(overlord.x + (overlord.width/2));
+//   var randomJunkFoodAttributes = new RandomJunkFood(overlordMouth);
+//   var junkFood = new JunkFood(randomJunkFoodAttributes);
+//   junkFood.ctx = ctx;
+//   junkFoods.push(junkFood);
+// }
